backend/requirements.txt
django==4.1.3
djangorestframework==3.13.1 
django-filter==2.4.0
django_cors_headers==3.13.0
psycopg2-binary==2.9.3
black==22.6.0
Pillow==9.2.0
ipython==8.4.0
isort[requirements]==4.3.21
pylint==2.14.4
watchdog[watchmedo]==0.10.3
django-model-utils==4.2.0
sentry-sdk==1.7.0
celery==5.2.7
redis==4.3.4
gunicorn==20.1.0
djangorestframework-simplejwt==5.2.0
funcy==1.17
ipdb==0.13.9
tzdata==2022.1
python-telegram-bot==13.14
python-telegram-bot-calendar==1.0.5
pytz==2022.6
tinkoff-investments==0.2.0b33
openai==0.27.2
numpy==1.24.2